Complete AoC2023 day 2
Took the opportunity to learn a bit about parsers in Rust, specifically
the "nom" crate. I suspect if I can get familiar with this crate, it
will make parsing future inputs slightly easier. Took me a while to
wrap my head around it though. It also took a while for me to figure out
the HashMap's "entry" API. Once I wrapped my head around it returning a
mutable reference to the existing value in the map, and found the
and_modify and or_insert functions, it became easier to use. The
examples in the reference docs were useful for this. I am a little
disappointed that there doesn't seem to be a nice "collect" function
which can handle the discrimination on duplicate keys, but the approach
I took here with fold and entry works out alright in this case. I am
consuming the Game object when I compute the score. I wonder if it is
possible to build this so that it uses references instead? That will
probably involve cloning some strings and integers around at some point.