Complete day 4 of Advent of Code 2023
Quite a nice one after finding day 3 difficult. Part 1 is fairly trivial
but part 2 puts an interesting spin on it which gave me pause. Putting
the card counts in a HashMap and relying on the fact that earlier card
counts cannot be affected by later cards lets us avoid having to process
many thousands of cards. We can just process each card once and multiply
the results by how many of that card we have.